refactor: split infer.rs into focused sub-modules
Extract Inferencer methods into dedicated files:
- expr.rs: infer_expr_inner + infer_binop (781 lines)
- pattern.rs: check_pattern + bind_pattern (185 lines)
- resolve.rs: resolve_type_expr*, fresh_subst_for (135 lines)

mod.rs reduced from 2238 to 1162 lines (~572 logic + ~590 tests).
All 41 tests passing.
